Działający backend
Klucze główne w bazie muszą być ustawione na autoinkremencjatcję. Żeby usuwać dane przez PostMana do URL trzeba wrzucić np. http://localhost:8080/api/v1/Ratownicy/1. Tak samo z putem. GET i POST działają normalnie bez /1 na końcu (/1 oznacza id tabeli)

+++ src/main/java/com/grspck/models/Sektory.java
@@ -1,15 +1,13 @@
 package com.grspck.models;
 
 import javax.persistence.*;
-import java.util.Collection;
-import java.util.List;
 
 @Entity(name = "sektory")
 public class Sektory
 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_sektora;
+    private Long id_sektora;
     private String nazwa;
 
     public Sektory()
@@ -17,11 +15,11 @@ public class Sektory
 
     }
 
-    public int getId_sektora() {
+    public Long getId_sektora() {
         return id_sektora;
     }
 
-    public void setId_sektora(int id_sektora) {
+    public void setId_sektora(Long id_sektora) {
         this.id_sektora = id_sektora;
     }
